chore: type consistency

+++ src/Governance.sol
@@ -146,7 +146,8 @@ contract Governance is Multicall, UserProxyFactory, ReentrancyGuard, IGovernance
 
         // 120 for timestamps = 2^32 * 1e18 | 2^32 * 1e26
         // 208 for voting power = 2^120 * 2^88
-        uint256 newOuterAverageAge;
+        // NOTE: 208 / X can go past u120!
+        uint208 newOuterAverageAge;
         if (_prevLQTYBalance <= _newLQTYBalance) {
             uint88 deltaLQTY = _newLQTYBalance - _prevLQTYBalance;
             uint208 prevVotes = uint208(_prevLQTYBalance) * uint208(prevOuterAverageAge);
